# Improved Python performance