Netřeba materializovat rok/měsíc přes strftime, je možné groupovat za běhu

Před:
```
$ python3 src/bitstamp.py 
Libraries loading time: 1.0624630451202393 s.
File loading time: 8.476861953735352 s.
Preprocessing time: 64.16597509384155 s.
Processing time: 1.7917855024337768 s.
```
Po:
```
$ python3 src/bitstamp.py 
Libraries loading time: 0.7324049472808838 s.
File loading time: 7.82699179649353 s.
Preprocessing time: 0.09575295448303223 s.
Processing time: 0.7583947896957397 s.
```